+++ tests/test_auth.py
@@ -39,6 +39,11 @@ from logfire.exceptions import LogfireConfigError
             'pylf_v2_eu_9f9ba85a-b759-4181-9527-d812e03f9f7f_0kYhc414Ys2FNDRdt5vFB05xFx5NjVcbcBMy4Kp6PH0W',
             'EU (https://logfire-eu.pydantic.dev) - pylf_v2_eu_9f9ba85a-b759-4181-9527-d812e03f9f7f_0kYhc****',
         ),
+        (
+            'https://logfire-eu.pydantic.dev',
+            'pylf_v3_eu_new-token-format',
+            'EU (https://logfire-eu.pydantic.dev) - new-t****',
+        ),
     ],
 )
 def test_user_token_str(base_url: str, token: str, expected: str) -> None: